+++ src/components/number/number.test.js
@@ -1,7 +1,6 @@
 import * as React from "react";
 import Number from "./number.component";
 import CypressMountWithProviders from "../../../cypress/support/component-helper/cypress-mount";
-
 import {
   getDataElementByValue,
   getElement,
@@ -12,10 +11,13 @@ import {
   commonInputPrefix,
   commonInputCharacterLimit,
 } from "../../../cypress/locators";
-
 import { verifyRequiredAsteriskForLabel } from "../../../cypress/support/component-helper/common-steps";
+import {
+  SIZE,
+  CHARACTERS,
+} from "../../../cypress/support/component-helper/constants";
 
-const testData = ["mp150ú¿¡üßä", "!@#$%^*()_+-=~[];:.,?{}&\"'<>"];
+const testData = [CHARACTERS.DIACRITICS, CHARACTERS.SPECIALCHARACTERS];
 
 const NumberInputComponent = ({ onChange, ...props }) => {
   const [state, setState] = React.useState("");
@@ -165,9 +167,9 @@ context("Tests for Number component", () => {
     });
 
     it.each([
-      ["small", "32px"],
-      ["medium", "40px"],
-      ["large", "48px"],
+      [SIZE.SMALL, "32px"],
+      [SIZE.MEDIUM, "40px"],
+      [SIZE.LARGE, "48px"],
     ])(
       "should use %s as size and render Number with %s as height",
       (size, height) => {
@@ -240,6 +242,27 @@ context("Tests for Number component", () => {
           .should("have.css", "width", inputRatio);
       }
     );
+  });
+
+  it.each(["10%", "30%", "50%", "80%", "100%"])(
+    "should check maxWidth as %s for Number component",
+    (maxWidth) => {
+      CypressMountWithProviders(<NumberInputComponent maxWidth={maxWidth} />);
+
+      getDataElementByValue("input")
+        .parent()
+        .parent()
+        .should("have.css", "max-width", maxWidth);
+    }
+  );
+
+  it("when maxWidth has no value it should render as 100%", () => {
+    CypressMountWithProviders(<NumberInputComponent maxWidth="" />);
+
+    getDataElementByValue("input")
+      .parent()
+      .parent()
+      .should("have.css", "max-width", "100%");
   });
 
   describe("check events for Number component", () => {